% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDMMA.R
\docType{data}
\name{BDMMA}
\alias{BDMMA}
\title{Bayesian Dirichlet--Multinomial approach for meta-analysis of metagenomic read counts}
\usage{
BDMMA(X, Y, batch, continuous, abundance_threshold = 5e-05, burn_in = 5000,
  sample_period = 5000, bFDR = 0.1)
}
\arguments{
\item{X}{A data.frame for covariates, including main effect variable and confounding variables.
The first column represents the main effect variable.}

\item{Y}{A data.frame for taxonomic read counts with the rows representing clades and columns representing taxa.}

\item{batch}{A numeric vector labeling the batch for each sample.}

\item{abundance_threshold}{The minimum abundance level for the taxa to be included (default value = 5e-05).}

\item{burn_in}{The length of burn in period before sampling the parameters (default value = 5,000).}

\item{sample_period}{The length of sampling period for estimating parameters' distribution (default value = 5,000)}

\item{bFDR}{The false discovery rate level to control (default value = 0.1).}
}
\value{
A list contains the selected taxa and summary of parameters included in the model.
\item{selected.taxa}{A list includes the selected taxa fesatures that are significantly associated
with the main effect variable.}
\item{parameter_summary}{A data.frame contains the mean and quantiles of the parameters included
in the model. Each row includes a parameter's distribution summary and the parameter name is
labeled in the first row. alpha_g: the baseline intercept of g-th taxon; betaj_g: the association strength between
the g-th taxon and j-th input variables; deltai_g: the batch effect parameter of batch i, taxon g;
L_g: the posterior selection probability of g-th taxon; p: the proportion of significantly associated
taxa; eta: the standard deviation of the spike distribution (in the spike-and-slab prior).}
}
\description{
Bayesian Dirichlet--Multinomial approach for meta-analysis of metagenomic read counts
}
\examples{
data(dat)
X <- dat$X
Y <- dat$Y
batch <- dat$batch
continuous <- dat$continuous
output <- BDMMA(X, Y, batch, continuous, burn_in = 3000, sample_period = 3000)
}
