% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbiome_dat.R
\docType{data}
\name{Microbiome_dat}
\alias{Microbiome_dat}
\title{Taxonomy Reads and Associated Phenotypes}
\format{SummarizedExperiment}
\usage{
Microbiome_dat
}
\description{
Simualated taxonomy read counts of 40 taxa and their associated
phenotypes.
}
\details{
The dataset contains the simualated taxonomy read counts from 80 samples,
where the samples come from 4 different batches and include both case and control
samples in each batch. For the detailed usuage, please see the package vignette.
}
\keyword{datasets}
