% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDMMA.R
\name{VBatch}
\alias{VBatch}
\title{Visualize batch effect with principal coordinate analysis}
\usage{
VBatch(Microbiome_dat, main_variable = NULL, method = "bray")
}
\arguments{
\item{Microbiome_dat}{A SummarizedExperiment object that includes the taxonomy read counts, phenotypes and batch labels.}

\item{main_variable}{Optional. A vector containing the main effect variable. Only for
categorical main effect variable.The function will generate a figure for each catagory.}

\item{method}{A string indicating which method should be used to calculate the distance matrix for
principal coordinate analysis.}
}
\value{
The function returns a list containing plot objects of principal coordinate analysis figures.
}
\description{
Visualize batch effect with principal coordinate analysis
}
\examples{
data(Microbiome_dat)
figure <- VBatch(Microbiome_dat, method = "bray")
print(figure)
}
